create randmaskCreater as module
